reservation test fails
