+++ Java/src/test/java/coltrain/WebTicketManagerTest.java
@@ -2,9 +2,18 @@ package coltrain;
 
 import org.junit.Test;
 
-public class WebTicketManagerTest {
-    @Test
-    public void reserve() {
+import static org.junit.Assert.assertEquals;
 
+public class WebTicketManagerTest {
+
+    @Test
+    public void reserve_givenZeroSeats_itShouldReturnNoSeats() {
+        WebTicketManager sut = new WebTicketManager();
+        String train = "express_2000";
+        int nbSeats = 0;
+
+
+        String reservation = sut.reserve(train, nbSeats);
+        assertEquals("{\"trainId\": \"express_2000\",\"bookingReference\": \"75bcd17\",\"seats\":[]}", reservation);
     }
 }
